+++ src/Type/AuditOperationType.php
@@ -11,7 +11,7 @@ namespace PrecisionSoft\Doctrine\Audit\Type;
 use PrecisionSoft\Doctrine\Audit\Dto\Operation;
 use PrecisionSoft\Doctrine\Type\Contract\AbstractEnumType;
 
-class AuditOperationType extends AbstractEnumType
+final class AuditOperationType extends AbstractEnumType
 {
     public function getValues(): array
     {